requirements.txt
anthropic>=0.40.0
requests>=2.31.0
cloudscraper>=1.2.71
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
numpy-financial>=1.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
flask>=3.0.0
